remove redundant plotting tests
