+++ autoarray/config/visualize/mat_wrap_2d.yaml
@@ -24,12 +24,14 @@ ArrayOverlay:        # wrapper for `plt.imshow()`: customize arrays overlaid.
 Contour:             # wrapper for `plt.contour()`: customize contours plotted on the figure.
   figure:
     colors: "k"
-    total_contours: 10
-    use_log10: true
+    total_contours: 10     # Number of contours to plot
+    use_log10: true        # If true, contours are plotted with log10 spacing, if False, linear spacing.
+    include_values: true   # If true, the values of the contours are plotted on the figure.
   subplot:
     colors: "k"
-    total_contours: 10
-    use_log10: true
+    total_contours: 10     # Number of contours to plot
+    use_log10: true        # If true, contours are plotted with log10 spacing, if False, linear spacing.
+    include_values: true   # If true, the values of the contours are plotted on the figure.
 BorderScatter:       # wrapper for `plt.scatter()`: customize the apperance of 2D borders.
   figure:
     c: r
