+++ test_autoarray/plot/wrap/one_d/test_axvline.py
@@ -1,26 +1,6 @@
 import autoarray.plot as aplt
 
 
-def test___from_config_or_via_manual_input():
-    line_plot = aplt.AXVLine()
-
-    assert line_plot.config_dict["ymin"] == 0.5
-
-    line_plot = aplt.AXVLine(ymin=0.7)
-
-    assert line_plot.config_dict["ymin"] == 0.7
-
-    line_plot = aplt.AXVLine()
-    line_plot.is_for_subplot = True
-
-    assert line_plot.config_dict["ymin"] == 0.6
-
-    line_plot = aplt.AXVLine(ymin=0.1)
-    line_plot.is_for_subplot = True
-
-    assert line_plot.config_dict["ymin"] == 0.1
-
-
 def test__plot_vertical_lines__works_for_reasonable_values():
     line = aplt.AXVLine(linewidth=2, linestyle="-", c="k")
