+++ test_autoarray/plot/wrap/two_d/test_grid_errorbar.py
@@ -2,30 +2,6 @@ import autoarray as aa
 import autoarray.plot as aplt
 
 
-def test__from_config_or_via_manual_input():
-    grid_errorbar = aplt.GridErrorbar()
-
-    assert grid_errorbar.config_dict["marker"] == "o"
-    assert grid_errorbar.config_dict["c"] == "k"
-
-    grid_errorbar = aplt.GridErrorbar(marker="x")
-
-    assert grid_errorbar.config_dict["marker"] == "x"
-    assert grid_errorbar.config_dict["c"] == "k"
-
-    grid_errorbar = aplt.GridErrorbar()
-    grid_errorbar.is_for_subplot = True
-
-    assert grid_errorbar.config_dict["marker"] == "."
-    assert grid_errorbar.config_dict["c"] == "b"
-
-    grid_errorbar = aplt.GridErrorbar(c=["r", "b"])
-    grid_errorbar.is_for_subplot = True
-
-    assert grid_errorbar.config_dict["marker"] == "."
-    assert grid_errorbar.config_dict["c"] == ["r", "b"]
-
-
 def test__errorbar_grid():
     errorbar = aplt.GridErrorbar(marker="x", c="k")
