+++ test_autoarray/plot/wrap/two_d/test_patcher.py
@@ -3,30 +3,6 @@ import autoarray.plot as aplt
 from matplotlib.patches import Ellipse
 
 
-def test__from_config_or_via_manual_input():
-    patch_overlay = aplt.PatchOverlay()
-
-    assert patch_overlay.config_dict["facecolor"] == None
-    assert patch_overlay.config_dict["edgecolor"] == "c"
-
-    patch_overlay = aplt.PatchOverlay(facecolor="r", edgecolor="g")
-
-    assert patch_overlay.config_dict["facecolor"] == "r"
-    assert patch_overlay.config_dict["edgecolor"] == "g"
-
-    patch_overlay = aplt.PatchOverlay()
-    patch_overlay.is_for_subplot = True
-
-    assert patch_overlay.config_dict["facecolor"] == None
-    assert patch_overlay.config_dict["edgecolor"] == "y"
-
-    patch_overlay = aplt.PatchOverlay(facecolor="b", edgecolor="p")
-    patch_overlay.is_for_subplot = True
-
-    assert patch_overlay.config_dict["facecolor"] == "b"
-    assert patch_overlay.config_dict["edgecolor"] == "p"
-
-
 def test__add_patches():
     patch_overlay = aplt.PatchOverlay(facecolor="c", edgecolor="none")
